Thanks to generics, no more cast is necessary when converting to a given class URL url = converter.convert(new URI("http://www.entropysoft.net"), URL.class);
This works also when converting to a parameterized class :
LinkedHashSet<String> linkedHashSet = converter.convert(source, LinkedHashSet.class, new Class[] { String.class });

+++ src/test/java/net/entropysoft/transmorph/DefaultFromStringConvertersTest.java
@@ -8,8 +8,7 @@ public class DefaultFromStringConvertersTest extends TestCase {
 		Transmorph transmorph = new Transmorph(
 				new DefaultFromStringConverters());
 
-		assertEquals(52.32, (Float) transmorph.convert("52.32", Float.TYPE),
-				0.001);
+		assertEquals(52.32, transmorph.convert("52.32", Float.TYPE), 0.001);
 		assertEquals(
 				DefaultFromStringConvertersTest.class,
 				(Class) transmorph
